Validaciones en Actividades Beneficas

+++ NFDao/Model/CharitableActivities.cs
@@ -13,7 +13,7 @@ namespace NFDao.Model
         public string name { get; set; }
         public string description { get; set; }
         public DateTime date { get; set; }
-        public double moneyRaising { get; set; }
+        public decimal moneyRaising { get; set; }
         public byte shelterId { get; set; }
         public int userId { get; set; }
         public CharitableActivities()
@@ -21,7 +21,7 @@ namespace NFDao.Model
             
         }
 
-        public CharitableActivities(int id, string name, string description, DateTime date, double moneyRaising, byte shelterId, int userId)
+        public CharitableActivities(int id, string name, string description, DateTime date, decimal moneyRaising, byte shelterId, int userId)
         {
             this.id = id;
             this.name = name;
@@ -32,7 +32,7 @@ namespace NFDao.Model
             this.userId = userId;
         }
 
-        public CharitableActivities(string name, string description, DateTime date, double moneyRaising, byte shelterId, int userId)
+        public CharitableActivities(string name, string description, DateTime date, decimal moneyRaising, byte shelterId, int userId)
         {
             this.name = name;
             this.description = description;
